Checkpoint: more site properties work, beginning of site class selection

+++ content/socialitePreferences.js
@@ -1,51 +1,74 @@
 Components.utils.import("resource://socialite/socialite.jsm");
 
-var siteListbox;
+var promptService = Components.classes["@mozilla.org/embedcomp/prompt-service;1"]
+                                       .getService(Components.interfaces.nsIPromptService); 
 
 var SocialiteSitePreferences = {
 
-  init: function() {
-    siteListbox = document.getElementById("socialiteSiteListbox"); 
-    for each (var site in Socialite.sites.byID) {
+  init: function SSPrefs_init() {
+    this.strings = document.getElementById("socialitePreferencesStrings");
+    this.siteListbox = document.getElementById("socialiteSiteListbox"); 
+    this.siteListbox.addSite = function(site) {
       var newItem = document.createElement("listitem");
       newItem.value = site;
       
       var siteCell = document.createElement("listcell");
-      siteCell.setAttribute("class", "listcell-iconic");
-      siteCell.setAttribute("label", site.siteName);
-      siteCell.setAttribute("image", site.getIconURI());
       newItem.appendChild(siteCell);
-      
       var urlCell = document.createElement("listcell");
-      urlCell.setAttribute("label", site.siteURL);
       newItem.appendChild(urlCell);
       
-      siteListbox.appendChild(newItem);
+      newItem.update = function() {
+        siteCell.setAttribute("class", "listcell-iconic");
+        siteCell.setAttribute("label", site.siteName);
+        siteCell.setAttribute("image", site.getIconURI());
+        urlCell.setAttribute("label", site.siteURL);
+      };
+      newItem.update();
+      
+      SocialiteSitePreferences.siteListbox.appendChild(newItem);
+    }
+    
+    for each (var site in Socialite.sites.byID) {
+      this.siteListbox.addSite(site);
     };
   },
 
-  siteAdd: function(event) {
-    var item = siteListbox.selectedItem;
+  siteAdd: function SSPrefs_siteAdd(event) {
+    var item = SocialiteSitePreferences.siteListbox.selectedItem;
     if (item) {
+      var dialog = document.documentElement.openSubDialog("chrome://socialite/content/socialiteSiteProperties.xul", "", {
+        isNewSite: true
+      });
+    }
+  },
+  
+  siteProperties: function SSPrefs_siteProperties(event) {
+    var item = SocialiteSitePreferences.siteListbox.selectedItem;
+    if (item) {
+      var site = item.value;
+      var dialog = document.documentElement.openSubDialog("chrome://socialite/content/socialiteSiteProperties.xul", "", {
+        isNewSite: false, 
+        site: site
+      });
+      item.update();
+    }
+  },
+  
+  siteRemove: function SSPrefs_siteRemove(event) {
+    var item = SocialiteSitePreferences.siteListbox.selectedItem;
+    if (item) {
+      var site = item.value;
       
-    }
-  },
-  
-  siteProperties: function(event) {
-    var item = siteListbox.selectedItem;
-    if (item) {
-      var site = item.value;
-      document.documentElement.openSubDialog("chrome://socialite/content/socialiteSiteProperties.xul", "", site)
-    }
-  },
-  
-  siteRemove: function(event) {
-    var item = siteListbox.selectedItem;
-    if (item) {
-      var site = item.value;
-      Socialite.sites.deleteSite(site);
-      Socialite.sites.saveConfiguredSites();
-      siteListbox.removeChild(item);
+      var confirmed = promptService.confirm(window, 
+          SocialiteSitePreferences.strings.getString("removeSiteConfirm.title"), 
+          SocialiteSitePreferences.strings.getFormattedString("removeSiteConfirm.message", [ site.siteName ])
+      );
+      
+      if (confirmed) {      
+        Socialite.sites.deleteSite(site);
+        Socialite.sites.saveConfiguredSites();
+        SocialiteSitePreferences.siteListbox.removeChild(item);
+      }
     }
   }
